requirements.txt
faiss-cpu
sentence-transformers
python-dotenv
groq
streamlit
